Add files via upload
This version of NCT uses cor_auto to compute polychoric correlations (in package qgraph of Sacha Epskamp). Note that the use of cor_auto in combination with NCT is not validated. When treating the data as ordinal, you assume that the data comes from a normal distribution and is thresholded into ordinal categories. When using cor_auto in NCT, you make the additional assumption that the thresholding is similar in both groups. It is the question whether you want to make this assumption. If you do not want to assume this, you might want to treat the data as normal in the ordinary NCT.